Fix displaying duplicate polygon token
