+++ data/src/main/kotlin/com/vultisig/wallet/data/usecases/chaintokens/CoinModificationStrategy.kt
@@ -1,0 +1,7 @@
+package com.vultisig.wallet.data.usecases.chaintokens
+
+import com.vultisig.wallet.data.models.Coin
+
+internal interface CoinModificationStrategy {
+    fun modify(item: Coin): Coin
+}
