+++ data/src/main/kotlin/com/vultisig/wallet/data/usecases/chaintokens/Constants.kt
@@ -1,0 +1,6 @@
+package com.vultisig.wallet.data.usecases.chaintokens
+
+internal const val LINK_TICKER = "LINK"
+
+// https://polygonscan.com/token/0x53e0bca35ec356bd5dddfebbd1fc0fd03fabad39
+internal const val BRIDGED_LINK_CONTRACT_ADDRESS = "0x53e0bca35ec356bd5dddfebbd1fc0fd03fabad39"
